# Conflicts:
#	src/main/java/com/example/Epic/Energy/Services/controller/AddressController.java
#	src/main/java/com/example/Epic/Energy/Services/entities/Address.java
#	src/main/java/com/example/Epic/Energy/Services/requests/AddressRequest.java
#	src/main/java/com/example/Epic/Energy/Services/services/AddressService.java

+++ src/main/java/com/example/Epic/Energy/Services/repositories/CustomerRepository.java
@@ -4,5 +4,5 @@ import com.example.Epic.Energy.Services.entities.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer>, PagingAndSortingRepository<Customer,Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long>, PagingAndSortingRepository<Customer,Long> {
 }
